+++ mockup/tests/pattern-structure-test.js
@@ -553,6 +553,7 @@ define([
         {
           model: {
             'Title': "Dummy Image",
+            'id': "dummy_page",
             'is_folderish': false,
             'portal_type': "Image",
             'getURL': 'http://nohost/dummy_image'
@@ -561,6 +562,7 @@ define([
         }, {
           model: {
             Title: "Dummy File",
+            id: "dummy_file",
             is_folderish: false,
             portal_type: "File",
             getURL: 'http://nohost/dummy_file'
@@ -569,6 +571,7 @@ define([
         }, {
           model: {
             Title: "Dummy Blob",
+            id: "dummy_blob",
             is_folderish: false,
             portal_type: "Blob",
             getURL: 'http://nohost/dummy_blob'
@@ -577,6 +580,7 @@ define([
         }, {
           model: {
             Title: "Dummy Document",
+            id: "dummy_document",
             is_folderish: false,
             portal_type: "Document",
             getURL: 'http://nohost/dummy_document'
@@ -610,6 +614,7 @@ define([
 
       var model = new Result({
           'Title': "Dummy Document",
+          'id': "dummy_document",
           'is_folderish': false,
           'portal_type': "Document",
           'getURL': 'http://nohost/dummy_image',
@@ -631,6 +636,7 @@ define([
 
       var model = new Result({
           'Title': "Dummy Document",
+          'id': "dummy_document",
           'is_folderish': false,
           'portal_type': "Document",
           'getURL': 'http://nohost/dummy_image',
@@ -670,6 +676,7 @@ define([
 
 			var collection = new ResultCollection([{
         'Title': 'Date Columns Test Document',
+        'id': 'date_columns_test_document',
         'is_folderish': false,
         'portal_type': 'Document',
         'getURL': 'http://nohost/doc',